@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "schiavinato-sharing"
-version = "0.5.0"
+version = "0.4.0"
 description = "Human-executable secret sharing for BIP39 mnemonics using GF(2053)"
 readme = "README.md"
 license = {text = "MIT"}
@@ -31,26 +31,27 @@ classifiers = [
     "Intended Audience :: Developers",
     "License :: OSI Approved :: MIT License",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
     "Topic :: Security :: Cryptography",
     "Topic :: Software Development :: Libraries :: Python Modules"
 ]
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 dependencies = [
     "mnemonic>=0.20",  # For BIP39 support
 ]
 
 [project.optional-dependencies]
 dev = [
-    "pytest>=7.4.0",
-    "pytest-cov>=4.1.0",
-    "mypy>=1.7.0",
-    "black>=23.12.0",
-    "ruff>=0.1.7",
-    "isort>=5.13.0",
+    # Pinned for deterministic CI and reproducible tooling behavior
+    # Keep pins compatible with the lowest supported runtime (Python 3.10)
+    "pytest==9.0.2",
+    "pytest-cov==7.0.0",
+    "mypy==1.19.1",
+    "black==26.1.0",
+    "ruff==0.14.14",
+    "isort==7.0.0",
 ]
 
 [project.urls]
@@ -71,7 +72,7 @@ python_functions = ["test_*"]
 addopts = "--verbose --cov=schiavinato_sharing --cov-report=term-missing --cov-fail-under=90"
 
 [tool.mypy]
-python_version = "3.9"
+python_version = "3.10"
 warn_return_any = true
 warn_unused_configs = true
 disallow_untyped_defs = true
@@ -88,12 +89,12 @@ strict_equality = true
 
 [tool.black]
 line-length = 100
-target-version = ['py39', 'py310', 'py311', 'py312']
+target-version = ['py310', 'py311', 'py312']
 include = '\.pyi?$'
 
 [tool.ruff]
 line-length = 100
-target-version = "py39"
+target-version = "py310"
 
 [tool.ruff.lint]
 select = [
