@@ -5,25 +5,40 @@ This module defines all data classes and types used by the library.
 """
 
 from dataclasses import dataclass, field
-from typing import List, Optional, Set
+from typing import TypedDict, cast
 
 
 @dataclass
 class Share:
     """
-    Represents a single Shamir share containing word shares, checksums, and Global Integrity Check (GIC).
+    Represents a single Shamir share containing word shares, checksums, and the
+    Global Integrity Check (GIC).
     """
+
     share_number: int
     """The X coordinate for this share (must be unique and non-zero)"""
-    
-    word_shares: List[int]
+
+    word_shares: list[int]
     """Array of word index shares (length = number of words in mnemonic)"""
-    
-    checksum_shares: List[int]
+
+    checksum_shares: list[int]
     """Array of row checksum shares (length = number of rows = wordCount / 3)"""
-    
+
     global_integrity_check_share: int
     """Global Integrity Check (GIC) verification number share"""
+
+
+RecoveryErrors = TypedDict(
+    "RecoveryErrors",
+    {
+        "row": list[int],
+        "global": bool,
+        "bip39": bool,
+        "generic": str | None,
+        "rowPathMismatch": list[int],
+        "globalPathMismatch": bool,
+    },
+)
 
 
 @dataclass
@@ -31,22 +46,27 @@ class RecoveryResult:
     """
     Result object returned by the recovery function.
     """
-    mnemonic: Optional[str] = None
+
+    mnemonic: str | None = None
     """The recovered mnemonic phrase (None if recovery failed)"""
-    
-    errors: dict = field(default_factory=lambda: {
-        'row': [],
-        'global': False,
-        'bip39': False,
-        'generic': None,
-        'rowPathMismatch': [],
-        'globalPathMismatch': False
-    })
+
+    errors: RecoveryErrors = field(
+        default_factory=lambda: cast(
+            RecoveryErrors,
+            {
+                "row": [],
+                "global": False,
+                "bip39": False,
+                "generic": None,
+                "rowPathMismatch": [],
+                "globalPathMismatch": False,
+            },
+        )
+    )
     """Detailed error information"""
-    
+
     success: bool = False
     """True if recovery was successful with no errors"""
-    
-    shares_with_invalid_checksums: Set[int] = field(default_factory=set)
-    """Set of share numbers that had invalid checksums (if any)"""
 
+    shares_with_invalid_checksums: set[int] = field(default_factory=set)
+    """Set of share numbers that had invalid checksums (if any)"""
