@@ -10,9 +10,6 @@ The prime 2053 was chosen because:
 - It allows for compact share representations
 """
 
-from typing import List, Tuple
-import functools
-
 # The prime modulus for the Galois field GF(2053)
 FIELD_PRIME = 2053
 
@@ -20,13 +17,13 @@ FIELD_PRIME = 2053
 def mod(value: int) -> int:
     """
     Reduces an integer into GF(2053) by applying modulo with wraparound for negatives.
-    
+
     Args:
         value: The integer value to reduce
-    
+
     Returns:
         The value reduced into the range [0, 2052]
-    
+
     Examples:
         >>> mod(2055)
         2
@@ -42,14 +39,14 @@ def mod(value: int) -> int:
 def mod_add(a: int, b: int) -> int:
     """
     Adds two field elements and returns the result modulo 2053.
-    
+
     Args:
         a: First field element
         b: Second field element
-    
+
     Returns:
         (a + b) mod 2053
-    
+
     Examples:
         >>> mod_add(2000, 100)
         47
@@ -62,14 +59,14 @@ def mod_add(a: int, b: int) -> int:
 def mod_sub(a: int, b: int) -> int:
     """
     Subtracts the second field element from the first inside GF(2053).
-    
+
     Args:
         a: First field element (minuend)
         b: Second field element (subtrahend)
-    
+
     Returns:
         (a - b) mod 2053
-    
+
     Examples:
         >>> mod_sub(100, 200)
         1953
@@ -82,14 +79,14 @@ def mod_sub(a: int, b: int) -> int:
 def mod_mul(a: int, b: int) -> int:
     """
     Multiplies two field elements and reduces the product into GF(2053).
-    
+
     Args:
         a: First field element
         b: Second field element
-    
+
     Returns:
         (a * b) mod 2053
-    
+
     Examples:
         >>> mod_mul(100, 50)
         841
@@ -103,16 +100,16 @@ def mod_inv(value: int) -> int:
     """
     Computes the multiplicative inverse of a non-zero field element.
     Uses the Extended Euclidean Algorithm.
-    
+
     Args:
         value: The field element to invert (must be non-zero)
-    
+
     Returns:
         The multiplicative inverse: value * result ≡ 1 (mod 2053)
-    
+
     Raises:
         ValueError: If value is 0 or not invertible
-    
+
     Examples:
         >>> mod_inv(2)
         1027
@@ -120,35 +117,35 @@ def mod_inv(value: int) -> int:
         2
     """
     val = mod(value)
-    
+
     if val == 0:
-        raise ValueError('Attempted to invert zero in GF(2053).')
-    
+        raise ValueError("Attempted to invert zero in GF(2053).")
+
     t, new_t = 0, 1
     r, new_r = FIELD_PRIME, val
-    
+
     while new_r != 0:
         quotient = r // new_r
         t, new_t = new_t, t - quotient * new_t
         r, new_r = new_r, r - quotient * new_r
-    
+
     if r > 1:
-        raise ValueError('Value is not invertible in GF(2053).')
-    
+        raise ValueError("Value is not invertible in GF(2053).")
+
     if t < 0:
         t += FIELD_PRIME
-    
+
     return t
 
 
 def mod_div(a: int, b: int) -> int:
     """
     Division in GF(2053): a / b = a * b^(-1).
-    
+
     Args:
         a: Dividend
         b: Divisor (must be non-zero)
-    
+
     Returns:
         (a / b) mod 2053
     """
@@ -158,11 +155,11 @@ def mod_div(a: int, b: int) -> int:
 def mod_pow(a: int, n: int) -> int:
     """
     Exponentiation in GF(2053).
-    
+
     Args:
         a: Base
         n: Exponent
-    
+
     Returns:
         a^n mod 2053
     """
@@ -172,13 +169,12 @@ def mod_pow(a: int, n: int) -> int:
 # For backward compatibility with validation code
 class GF2053:
     """Legacy class-based interface for GF(2053) arithmetic."""
-    
+
     PRIME = FIELD_PRIME
-    
+
     add = staticmethod(mod_add)
     sub = staticmethod(mod_sub)
     mul = staticmethod(mod_mul)
     inv = staticmethod(mod_inv)
     div = staticmethod(mod_div)
     pow = staticmethod(mod_pow)
-
