@@ -2,10 +2,11 @@ import pytest
 
 from schiavinato_sharing import (
     generate_valid_mnemonic,
-    mnemonic_to_indices,
     indices_to_mnemonic,
+    mnemonic_to_indices,
     parse_input,
 )
+from schiavinato_sharing.seed import validate_bip39_mnemonic
 
 
 def test_mnemonic_index_roundtrip():
@@ -37,3 +38,32 @@ def test_parse_input_rejects_unknown_or_out_of_range():
         parse_input("3000")  # out of range (valid range is 1-2048)
     with pytest.raises(ValueError):
         parse_input("0")  # 0 is no longer valid (1-based indexing)
+
+
+def test_validate_bip39_mnemonic_accepts_known_good_and_rejects_bad_inputs():
+    # Known BIP39 test mnemonic (12 words)
+    mnemonic = (
+        "abandon abandon abandon abandon abandon abandon abandon abandon "
+        "abandon abandon abandon about"
+    )
+    assert validate_bip39_mnemonic(mnemonic) is True
+
+    # Non-string inputs are rejected
+    assert validate_bip39_mnemonic(None) is False  # type: ignore[arg-type]
+
+    # Invalid length is rejected (11 words)
+    assert validate_bip39_mnemonic("abandon " * 10 + "about") is False
+
+    # Unknown word is rejected
+    bad = mnemonic.replace("about", "notaword")
+    assert validate_bip39_mnemonic(bad) is False
+
+
+def test_generate_valid_mnemonic_rejects_unsupported_word_count():
+    with pytest.raises(ValueError):
+        generate_valid_mnemonic(15)
+
+
+def test_indices_to_mnemonic_rejects_out_of_range_indices():
+    with pytest.raises(ValueError):
+        indices_to_mnemonic([0])
